% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rating_table.R
\name{add_observed_experience}
\alias{add_observed_experience}
\title{Deprecated alias for \code{add_portfolio_experience()}}
\usage{
add_observed_experience(...)
}
\arguments{
\item{...}{Unused.}
}
\value{
See \code{\link[=add_portfolio_experience]{add_portfolio_experience()}}.
}
\description{
\code{add_observed_experience()} is deprecated. Use
\code{\link[=add_portfolio_experience]{add_portfolio_experience()}} instead.
}
\author{
Martin Haringa
}
\keyword{internal}
