% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_restrictions_df}
\alias{add_restrictions_df}
\title{Join restricted data to model data}
\usage{
add_restrictions_df(model_data, restrictions_df)
}
\arguments{
\item{model_data}{data.frame with original model data}

\item{restrictions_df}{data.frame with two columns, column 1 must include
the levels of the risk factor, and column 2 must include the new restricted
coefficients.}
}
\description{
Join restricted data to model data
}
\keyword{internal}
