% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_riskfactor.R
\name{as.data.frame.fitgam}
\alias{as.data.frame.fitgam}
\title{Coerce fitgam objects to a data frame}
\usage{
\method{as.data.frame}{fitgam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fitgam"}.}

\item{...}{Further arguments passed to \code{\link[=as.data.frame]{as.data.frame()}}.}
}
\value{
A \code{data.frame} containing the predictions.
}
\description{
Extracts the prediction component of a \code{fitgam} object and returns it
as a data frame.
}
