% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_bootstrap_rmse.R
\name{as.vector.bootstrap_performance}
\alias{as.vector.bootstrap_performance}
\title{Coerce bootstrap_performance objects to a vector}
\usage{
\method{as.vector}{bootstrap_performance}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bootstrap_performance"}.}

\item{...}{Further arguments passed to \code{\link[=as.vector]{as.vector()}}.}
}
\value{
A numeric vector with the bootstrap RMSE values.
}
\description{
Extracts the bootstrap RMSE values as a numeric vector.
}
