% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_construct_tariff_classes.R
\name{as.vector.constructtariffclasses}
\alias{as.vector.constructtariffclasses}
\title{Coerce constructtariffclasses to a vector}
\usage{
\method{as.vector}{constructtariffclasses}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"constructtariffclasses"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
A numeric vector with the split points of the tariff classes.
}
\description{
Extracts the tariff class splits as a numeric vector.
}
