% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{autoplot.check_normality}
\alias{autoplot.check_normality}
\title{Automatically create a ggplot for objects obtained from check_normality()}
\usage{
\method{autoplot}{check_normality}(object, type = c("qq", "density"), data = NULL, ...)
}
\arguments{
\item{object}{check_normality object produced by \code{check_normality()}}

\item{type}{character vector, "qq" (default) or "density"}

\item{data}{data (defaults to NULL)}

\item{...}{other plotting parameters to affect the plot}
}
\value{
a ggplot object
}
\description{
Takes an object produced by \code{check_normality()}, and plots the available input.
}
\author{
Martin Haringa
}
