% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_loss.R
\name{autoplot.excess_loss_allocation}
\alias{autoplot.excess_loss_allocation}
\title{Plot an excess-loss allocation}
\usage{
\method{autoplot}{excess_loss_allocation}(
  object,
  y = c("allocated_loading", "allocated_excess_loss", "credibility"),
  ...
)
}
\arguments{
\item{object}{An object returned by \code{\link[=allocate_excess_loss]{allocate_excess_loss()}}.}

\item{y}{Character. Measure to plot on the y-axis.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise the allocated excess loading, allocated excess loss or credibility
by allocation group.
}
\author{
Martin Haringa
}
