% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tariff_classes.R
\name{autoplot.insurancerating}
\alias{autoplot.insurancerating}
\title{Automatically create a ggplot for objects obtained from construct_tariff_classes()}
\usage{
\method{autoplot}{insurancerating}(gam_x, conf.int = FALSE,
  clusters = TRUE, color_gam = "steelblue", color_splits = "grey50",
  xstep = 10)
}
\arguments{
\item{gam_x}{A fitted gam object as produced by \code{construct_tariff_classes()}.}

\item{conf.int}{Determines whether 95\% confidence intervals will be plotted. The default is \code{conf.int = FALSE}.}

\item{clusters}{Numerical vector with splits as produced by \code{construct_tariff_classes()}.}

\item{color_gam}{A color can be specified either by name (e.g.: "red") or by hexadecimal code (e.g. : "#FF1234") (default is "steelblue").}

\item{color_splits}{change the color of the splits in the graph ("grey50" is default).}

\item{xstep}{Set step size for horizontal axis (default is \code{10}).}
}
\value{
Produces plot showing the smooth components of a fitted GAM.
}
\description{
Takes a fitted gam object produced by \code{construct_tariff_classes()} and plots the component smooth functions that make it up,
on the scale of the predicted number of claims.
}
\examples{
library(ggplot2)
x <- construct_tariff_classes(MTPL, nclaims, age_policyholder, exposure)
autoplot(x)

}
\author{
Martin Haringa
}
