% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{autoplot.smooth}
\alias{autoplot.smooth}
\title{Automatically create a ggplot for objects obtained from add_smoothing()}
\usage{
\method{autoplot}{smooth}(object, ...)
}
\arguments{
\item{object}{Object produced by \code{add_smoothing()}.}

\item{...}{Additional plotting arguments passed to ggplot2 geoms.}
}
\value{
A \code{ggplot2} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Takes an object produced by \code{add_smoothing()} and creates a plot comparing
the original model fit, the new clustered values, and the resulting smooth.
}
\author{
Martin Haringa
}
