% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{big.ref}
\alias{big.ref}
\title{Relevel factor to biggest group}
\usage{
big.ref(x, exposure)
}
\arguments{
\item{x}{A vector}

\item{exposure}{A vector with exposures}
}
\value{
A factor with releveled levels
}
\description{
Argument for setting the reference group to the biggest group in linear regression.
}
\examples{
mtcars2 <- mtcars1 \%>\%
mutate_if(is.factor, funs(big.ref(., qsec)))
}
