% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biggest_reference}
\alias{biggest_reference}
\title{Deprecated alias for \code{set_reference_level()}}
\usage{
biggest_reference(x, weight)
}
\arguments{
\item{x}{A factor.}

\item{weight}{A numeric vector of the same length as \code{x}.}
}
\value{
See \code{\link[=set_reference_level]{set_reference_level()}}.
}
\description{
\code{biggest_reference()} is deprecated as of version 0.9.0. Use
\code{\link[=set_reference_level]{set_reference_level()}} instead.
}
\keyword{internal}
