% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_overdispersion.R
\name{check_overdispersion}
\alias{check_overdispersion}
\title{Check overdispersion of a Poisson claim frequency model}
\usage{
check_overdispersion(object)
}
\arguments{
\item{object}{A fitted model of class \code{"glm"} with family Poisson.}
}
\value{
An object of class \code{"overdispersion_check"} and \code{"overdispersion"},
which is a list with elements:
\describe{
\item{pearson_chisq}{Pearson's chi-squared statistic.}
\item{dispersion_ratio}{Dispersion ratio, calculated as Pearson's
chi-squared statistic divided by residual degrees of freedom.}
\item{residual_df}{Residual degrees of freedom.}
\item{p_value}{P-value from the chi-squared test.}
}
For backwards compatibility the object also contains the aliases \code{chisq},
\code{ratio}, \code{rdf}, and \code{p}.
}
\description{
Tests whether a fitted Poisson GLM shows overdispersion using Pearson's
chi-squared statistic.
}
\details{
In Poisson claim frequency models, the variance is assumed to be equal to the
mean. A dispersion ratio above 1 indicates that the observed variation is
larger than expected under that assumption. In pricing work this can be a
useful diagnostic signal for omitted heterogeneity, clustering, outliers, or
model misspecification. It does not automatically mean that the model is
unusable.
\itemize{
\item A dispersion ratio close to 1 is broadly consistent with the Poisson
variance assumption.
\item A dispersion ratio above 1 suggests overdispersion.
\item A p-value below 0.05 indicates statistically significant overdispersion.
}
}
\examples{
x <- glm(nclaims ~ area, offset = log(exposure),
         family = poisson(), data = MTPL2)
check_overdispersion(x)

}
\references{
Bolker B. et al. (2017). \href{http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html}{GLMM FAQ}
See also: \code{performance::check_overdispersion()}.
}
\author{
Martin Haringa
}
