% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_classify.R
\name{fisher}
\alias{fisher}
\title{Deprecated alias for \code{fisher_classify()}}
\usage{
fisher(x, n = 7, diglab = 2)
}
\arguments{
\item{x}{Numeric vector to classify.}

\item{n}{Number of classes.}

\item{diglab}{Deprecated. Use \code{dig.lab} in \code{\link[=fisher_classify]{fisher_classify()}} instead.}
}
\value{
See \code{\link[=fisher_classify]{fisher_classify()}}.
}
\description{
\code{fisher()} is deprecated as of version 0.8.0.
}
\keyword{internal}
