% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_classify.R
\name{fisher_classify}
\alias{fisher_classify}
\alias{fisher}
\title{Fisher's natural breaks classification}
\usage{
fisher_classify(x, n = 7, dig.lab = NULL, diglab = NULL)

fisher(x, n = 7, diglab = 2)
}
\arguments{
\item{x}{A numeric vector to be classified.}

\item{n}{Integer. Number of classes to generate (default = 7).}

\item{dig.lab}{Integer. Number of significant digits to use for interval
labels (default = 2).}

\item{diglab}{Deprecated. Use \code{dig.lab} instead.}
}
\value{
A factor indicating the interval to which each element of \code{x}
belongs.
}
\description{
Classifies a continuous numeric vector into intervals using
Fisher-Jenks natural breaks. Useful for choropleth mapping or
other applications where grouped ranges are required.

\code{fisher()} is deprecated as of version 0.8.0.
Please use \code{\link[=fisher_classify]{fisher_classify()}} instead.
}
\details{
The \code{"fisher"} style uses the algorithm proposed by Fisher (1958), commonly
referred to as the Fisher-Jenks algorithm. This function is a thin wrapper
around \code{\link[classInt:classIntervals]{classInt::classIntervals()}}.

The argument \code{diglab} is deprecated and will be removed in a future version.
}
\examples{
set.seed(1)
x <- rnorm(100)
fisher_classify(x, n = 5)

}
\references{
Bivand, R. (2018). \emph{classInt: Choose Univariate Class Intervals}.
R package version 0.2-3. \url{https://CRAN.R-project.org/package=classInt}

Fisher, W. D. (1958). \emph{On grouping for maximum homogeneity}.
Journal of the American Statistical Association, 53, pp. 789–798.
doi:10.1080/01621459.1958.10501479
}
\author{
Martin Haringa
}
