% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_cut.R
\name{gam_cut}
\alias{gam_cut}
\title{GAM cut}
\usage{
gam_cut(data, nclaims, x, exposure, cp = 0, color_splits = "red",
  color_gam = "steelblue", show_splits = TRUE)
}
\arguments{
\item{exposure}{}
}
\description{
GAM cut
}
\author{
Martin Haringa
}
