% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_frequency.R
\name{gam_frequency}
\alias{gam_frequency}
\title{Fitting GAM for continuous variables in insurance rating (frequency)}
\usage{
gam_frequency(data, nclaims, x, exposure, accuracy_x = 1)
}
\arguments{
\item{data}{data.frame of an insurance portfolio.}

\item{nclaims}{column in \code{data} with number of claims.}

\item{x}{column in \code{data} with continuous risk factor.}

\item{exposure}{column in \code{data} with exposure.}

\item{accuracy_x}{round elements in column \code{x} to multiple of number \code{accuracy_x} (default value is 1). GAMs can be slow for fitting models with large
datasets. Therefore, an argument is added such that elements in the continuous risk factor are rounded and aggregated to a multiple of number \code{accuracy_x}.
Setting this argument equal to 1 results in a large speed enhancement. Setting this argument equal to 0 results in using the original values for the
continuous risk factor.}
}
\value{
The function returns an object of class "gam".
}
\description{
The function provides an interface to finding class intervals for continuous numerical variables, for example to bin the continuous factors
such that categorical risk factors result which capture the effect of the covariate on the response in an accurate way,
while being easy to use in a generalized linear model (GLM).
}
\details{
A Poisson distribution is assumed for the number of claims. The logarithm of the exposure is
included in the model as an offset, such that the expected number of claims is proportional to exposure.
}
\examples{
gam_frequency(MTPL, nclaims, age_policyholder, exposure)
}
\references{
Antonio, K. and Valdez, E. A. (2012). Statistical concepts of a priori and a posteriori risk classification in insurance. Advances in Statistical Analysis, 96(2):187–224.
}
\author{
Martin Haringa
}
