% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_frequency_clusters.R
\name{gam_frequency_clusters}
\alias{gam_frequency_clusters}
\title{Breaks for continuous variables in insurance rating}
\usage{
gam_frequency_clusters(gam_x, cp = 0)
}
\arguments{
\item{gam_x}{A fitted gam object as produced by \code{gam_frequency()}.}

\item{cp}{complexity parameter. Any split that does not decrease the overall lack of fit by a factor of cp is not attempted.
For instance, with anova splitting, this means that the overall R-squared must increase by \code{cp} at each step.
The main role of this parameter is to save computing time by pruning off splits that are obviously not worthwhile.
Essentially,the user informs the program that any split which does not improve the fit by \code{cp} will likely be pruned off
by cross-validation, and that hence the program need not pursue it (\code{cp} = 0 is default).}
}
\value{
Numerical vector with splits.
}
\description{
The function provides an interface to finding class intervals for continuous numerical variables, for example to bin the continuous factors
such that categorical risk factors result which capture the effect of the covariate on the response in an accurate way,
while being easy to use in a generalized linear model (GLM).
}
\details{
Bins are created where consecutive values of a continuous risk factor are grouped together. Regression trees
are used as a technique to perform the binning. Here regression trees from the \code{rpart} package are used.
}
\examples{
gam_fit <- gam_frequency(MTPL, nclaims, age_policyholder, exposure)
gam_frequency_clusters(gam_fit)
}
\references{
Antonio, K. and Valdez, E. A. (2012). Statistical concepts of a priori and a posteriori risk classification in insurance. Advances in Statistical Analysis, 96(2):187–224.
}
\author{
Martin Haringa
}
