% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_frequency_clusters_plot.R
\name{gam_frequency_clusters_plot}
\alias{gam_frequency_clusters_plot}
\title{GAM frequency plot (with class intervals added)}
\usage{
gam_frequency_clusters_plot(gam_x, clusters, xlab = "x", xstep = 10,
  color_splits = "grey50", color_gam = "steelblue")
}
\arguments{
\item{gam_x}{A fitted gam object as produced by \code{gam_frequency()}.}

\item{clusters}{Numerical vector with splits as produced by \code{gam_frequency_clusters()}.}

\item{xlab}{If supplied then this will be used as the x label.}

\item{xstep}{Set step size for horizontal axis (default is 10).}

\item{color_splits}{change the color of the splits in the graph ("grey50" is default).}

\item{color_gam}{change the color of the line in the graph ("steelblue" is default).}
}
\value{
Produces plot showing the smooth components of a fitted GAM. Class intervals for continuous numerical variables, binned by decision trees, are shown.
}
\description{
Takes a fitted gam object produced by \code{gam_frequency()}, and a numerical vector with splits as produced by \code{gam_frequency_clusters()}.
}
\examples{
gam_fit <- gam_frequency(MTPL, nclaims, age_policyholder, exposure)
clusters <- gam_frequency_clusters(gam_fit)
gam_frequency_clusters_plot(gam_fit, clusters)
}
\author{
Martin Haringa
}
