% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_frequency_plot.R
\name{gam_frequency_plot}
\alias{gam_frequency_plot}
\title{GAM frequency plot}
\usage{
gam_frequency_plot(gam_x, xlab = "x", xstep = 10,
  color_gam = "steelblue")
}
\arguments{
\item{gam_x}{A fitted gam object as produced by \code{gam_frequency()}.}

\item{xlab}{If supplied then this will be used as the x label.}

\item{xstep}{Set step size for horizontal axis (default is \code{10}).}

\item{color_gam}{A color can be specified either by name (e.g.: "red") or by hexadecimal code (e.g. : "#FF1234") (default is "steelblue").}
}
\value{
Produces plot showing the smooth components of a fitted GAM.
}
\description{
Takes a fitted gam object produced by \code{gam_frequency()} and plots the component smooth functions that make it up, on the scale of the predicted number of claims.
}
\examples{
gam_fit <- gam_frequency(MTPL, nclaims, age_policyholder, exposure)
gam_frequency_plot(gam_fit)
}
\author{
Martin Haringa
}
