% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{get_data}
\alias{get_data}
\title{Get data from refitted Generalized Linear Model}
\usage{
get_data(x)
}
\arguments{
\item{x}{Object of class refitsmooth or of class refitrestricted}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{get_data()} is used to get data from refitted generalized linear models, and must be
preceded by \code{refit_glm()}.
}
\author{
Martin Haringa
}
