% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{get_offset}
\alias{get_offset}
\title{Get offset from model object}
\usage{
get_offset(model)
}
\arguments{
\item{model}{Must be of class glm}
}
\value{
Character string with offset term
}
\description{
Get offset from model object
}
\keyword{internal}
