% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_stars}
\alias{make_stars}
\title{Convert p-values into significance stars}
\usage{
make_stars(pval)
}
\arguments{
\item{pval}{Numeric vector of p-values.}
}
\value{
Character vector of the same length as \code{pval},
containing \code{"***"}, \code{"**"}, \code{"*"}, \code{"."}, or \code{""}.
}
\description{
Convert p-values into significance stars
}
\keyword{internal}
