% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_data.R
\name{model_data}
\alias{model_data}
\title{Deprecated alias for \code{extract_model_data()}}
\usage{
model_data(x)
}
\arguments{
\item{x}{An object of class \code{"glm"}, \code{"refitsmooth"}, or \code{"refitrestricted"}.}
}
\value{
See \code{\link[=extract_model_data]{extract_model_data()}}.
}
\description{
\code{model_data()} is deprecated in favour of \code{\link[=extract_model_data]{extract_model_data()}}.
}
\keyword{internal}
