% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_period_to_months.R
\name{period_to_months}
\alias{period_to_months}
\title{Deprecated alias for \code{split_periods_to_months()}}
\usage{
period_to_months(df, begin, end, ...)
}
\arguments{
\item{begin}{Deprecated NSE argument. Use \code{period_start} instead.}

\item{end}{Deprecated NSE argument. Use \code{period_end} instead.}

\item{...}{Deprecated NSE columns to prorate. Use \code{prorate_cols} instead.}
}
\value{
See \code{\link[=split_periods_to_months]{split_periods_to_months()}}.
}
\description{
\code{period_to_months()} is deprecated as of version 0.8.0. Use
\code{\link[=split_periods_to_months]{split_periods_to_months()}} instead.
}
\keyword{internal}
