% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_construct_tariff_classes.R
\name{print.constructtariffclasses}
\alias{print.constructtariffclasses}
\title{Print method for constructtariffclasses objects}
\usage{
\method{print}{constructtariffclasses}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"constructtariffclasses"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays the tariff class splits of an object created by
\code{\link[=construct_tariff_classes]{construct_tariff_classes()}}.
}
