% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_loss.R
\name{print.excess_loss_vector}
\alias{print.excess_loss_vector}
\title{Print an excess-loss vector}
\usage{
\method{print}{excess_loss_vector}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"excess_loss_vector"}.}

\item{...}{Reserved for future extensions.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Compact print method for objects returned by \code{\link[=calculate_excess_loss]{calculate_excess_loss()}}.
}
\author{
Martin Haringa
}
\keyword{internal}
