% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_riskfactor.R
\name{print.fitgam}
\alias{print.fitgam}
\title{Print method for fitgam objects}
\usage{
\method{print}{fitgam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fitgam"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the prediction component of a \code{fitgam} object created by
\code{\link[=riskfactor_gam]{riskfactor_gam()}}.
}
