% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{print.model_performance}
\alias{print.model_performance}
\title{Print method for model_performance objects}
\usage{
\method{print}{model_performance}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"model_performance"}.}

\item{digits}{Number of digits to round numeric columns. Default = 3.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Nicely formats and prints the results of \code{\link[=model_performance]{model_performance()}}, including
rounded numeric values, without requiring external packages.
}
