% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{print.restricted}
\alias{print.restricted}
\title{Print for object of class restricted}
\usage{
\method{print}{restricted}(x, ...)
}
\arguments{
\item{x}{Object of class restricted}

\item{...}{other plotting parameters to affect the output}
}
\value{
Print object
}
\description{
Print for object of class restricted
}
