% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rating_table.R
\name{rating_factors2}
\alias{rating_factors2}
\title{Include reference group in regression output}
\usage{
rating_factors2(
  model,
  model_data = NULL,
  exposure = TRUE,
  exposure_name = NULL,
  colname = "estimate",
  exponentiate = TRUE,
  round_exposure = 0
)
}
\arguments{
\item{model}{glm object produced by \code{glm()}}

\item{model_data}{Optional data.frame used to create glm object. If \code{NULL},
the function tries to use \code{model$data}.}

\item{exposure}{Logical or character. If \code{TRUE} (default), exposure is added
if it can be inferred from the model. If \code{FALSE}, no exposure is added.
If a character string is supplied, it is interpreted as the exposure column
name.}

\item{exposure_name}{Optional name for the exposure column in the output.}

\item{colname}{name of coefficient column}

\item{exponentiate}{logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to TRUE.}

\item{round_exposure}{number of digits for exposure (defaults to 0)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Legacy interface. Prefer \code{\link[=rating_table]{rating_table()}} for fitted models in the new
workflow, but this function remains available.
}
