% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_reduce.R
\name{reduce}
\alias{reduce}
\title{Deprecated alias for \code{merge_date_ranges()}}
\usage{
reduce(df, begin, end, ..., agg_cols = NULL, agg = "sum", min.gapwidth = 5)
}
\arguments{
\item{begin}{Deprecated NSE argument. Use \code{period_start} instead.}

\item{end}{Deprecated NSE argument. Use \code{period_end} instead.}

\item{...}{Deprecated NSE grouping columns. Use \code{group_by} instead.}

\item{agg_cols}{Deprecated NSE argument. Use \code{aggregate_cols} instead.}

\item{agg}{Deprecated. Use \code{aggregate_fun} instead.}

\item{min.gapwidth}{Deprecated. Use \code{merge_gap_days} instead.}
}
\value{
See \code{\link[=merge_date_ranges]{merge_date_ranges()}}.
}
\description{
\code{reduce()} is deprecated as of version 0.8.0. Use
\code{\link[=merge_date_ranges]{merge_date_ranges()}} instead.
}
\keyword{internal}
