% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit_glm}
\alias{refit_glm}
\title{Deprecated refit wrapper}
\usage{
refit_glm(x, intercept_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{rating_refinement}, \code{restricted} or \code{smooth}.}

\item{intercept_only}{Logical.}

\item{...}{Other arguments.}
}
\value{
Object of class \code{glm}.
}
\description{
\code{refit_glm()} is deprecated as of version 0.9.0. Use \code{\link[=refit]{refit()}} instead.
}
\keyword{internal}
