% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit_glm.restricted}
\alias{refit_glm.restricted}
\title{Refit GLM}
\usage{
refit_glm.restricted(x)
}
\arguments{
\item{x}{Object of class restricted}
}
\value{
Object of class GLM
}
\description{
Refit GLM model
}
