% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{remove_offset_formula}
\alias{remove_offset_formula}
\title{Remove offset term from formula}
\usage{
remove_offset_formula(formula)
}
\arguments{
\item{formula}{Formula of class formula}
}
\description{
Remove offset term from formula
}
\keyword{internal}
