% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{smooth_coef}
\alias{smooth_coef}
\title{Deprecated smoothing helper}
\usage{
smooth_coef(
  model,
  x_cut,
  x_org,
  degree = NULL,
  breaks = NULL,
  smoothing = "spline",
  k = NULL,
  weights = NULL
)
}
\arguments{
\item{model}{A fitted model object.}

\item{x_cut}{Deprecated model variable used in the GLM.}

\item{x_org}{Deprecated source variable used to fit the smoothing curve.}

\item{degree}{Deprecated polynomial degree.}

\item{breaks}{Deprecated smoothing break points.}

\item{smoothing}{Deprecated smoothing type.}

\item{k}{Deprecated spline basis dimension.}

\item{weights}{Deprecated weights column.}
}
\value{
A legacy smooth object. New code should use \code{\link[=prepare_refinement]{prepare_refinement()}},
\code{\link[=add_smoothing]{add_smoothing()}}, and \code{\link[=refit]{refit()}}.
}
\description{
\code{smooth_coef()} is deprecated as of version 0.9.0. Use
\code{\link[=add_smoothing]{add_smoothing()}} instead.

\preformatted{
prepare_refinement(model) |>
  add_smoothing(...) |>
  refit()
}
}
\seealso{
\code{\link[=add_smoothing]{add_smoothing()}}, \code{\link[=prepare_refinement]{prepare_refinement()}}, \code{\link[=refit]{refit()}}
}
\keyword{internal}
