% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_6exposure.R
\name{univariate_exposure}
\alias{univariate_exposure}
\title{Univariate exposure}
\usage{
univariate_exposure(df, x, exposure = exposure)
}
\arguments{
\item{df}{data.frame with insurance portfolio}

\item{x}{column in \code{df} with risk factor}

\item{exposure}{column in \code{df} with exposure}
}
\value{
An list of class \code{univ_lossratio} with components
\item{df}{data frame with average premium}
\item{xvar}{name of column in df with risk factor}
\item{exposure}{name of column in df with exposure}
}
\description{
Exposure for discrete risk factors in an insurance portfolio.
}
\examples{
univariate_exposure(MTPL2, area, exposure)
}
\author{
Martin Haringa
}
