% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_refinement.R
\name{update_formula_add}
\alias{update_formula_add}
\title{Create new offset-term and new formula}
\usage{
update_formula_add(offset_term, fm_no_offset, add_term)
}
\arguments{
\item{offset_term}{String obtained from get_offset()}

\item{fm_no_offset}{Obtained from remove_offset_formula()}

\item{add_term}{Name of restricted risk factor to add}
}
\description{
Create new offset-term and new formula
}
\keyword{internal}
