% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{update_formula_remove}
\alias{update_formula_remove}
\title{Remove risk factor from formula}
\usage{
update_formula_remove(fm, remove_term)
}
\arguments{
\item{fm}{Formula of class formula}

\item{remove_term}{Risk factor to remove}
}
\description{
Remove risk factor from formula
}
\keyword{internal}
