% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{update_glm}
\alias{update_glm}
\title{Deprecated alias for \code{refit_glm()}}
\usage{
update_glm(x, intercept_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{rating_refinement}, \code{restricted} or \code{smooth}.}

\item{intercept_only}{Logical.}

\item{...}{Other arguments.}
}
\value{
See \code{\link[=refit_glm]{refit_glm()}}.
}
\description{
\code{update_glm()} is deprecated as of version 0.8.0. Use \code{\link[=refit]{refit()}} for the new
refinement workflow.
}
\keyword{internal}
