% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_relativities}
\alias{add_relativities}
\title{Add expert-based relativities to a refinement workflow}
\usage{
add_relativities(
  model,
  model_variable,
  split_variable,
  relativities,
  exposure,
  normalize = TRUE
)
}
\arguments{
\item{model}{Object of class \code{rating_refinement}, usually created with
\code{\link[=prepare_refinement]{prepare_refinement()}}.}

\item{model_variable}{Character string. Existing variable in the GLM. Levels
of this variable can be split into more detailed groups.}

\item{split_variable}{Character string. More granular portfolio variable that
defines the split groups.}

\item{relativities}{Named list of data.frames, usually created with
\code{\link[=relativities_list]{relativities_list()}} and \code{\link[=split_level]{split_level()}}.}

\item{exposure}{Character string. Exposure column.}

\item{normalize}{Logical.}
}
\value{
Object of class \code{rating_refinement}.
}
\description{
Adds a relativities step to a \code{rating_refinement} object.
}
\details{
\code{add_relativities()} is only available within the new refinement workflow
and should be used in combination with \code{prepare_refinement()} and \code{refit()}.
}
