% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_relativities}
\alias{add_relativities}
\title{Add expert-based relativities to a refinement workflow}
\usage{
add_relativities(
  model,
  risk_factor,
  risk_factor_split,
  relativities,
  exposure,
  normalize = TRUE
)
}
\arguments{
\item{model}{Object of class \code{rating_refinement}.}

\item{risk_factor}{Character string. Name of existing risk factor in the model.}

\item{risk_factor_split}{Character string. More granular split column.}

\item{relativities}{Named list of data.frames.}

\item{exposure}{Character string. Exposure column.}

\item{normalize}{Logical.}
}
\value{
Object of class \code{rating_refinement}.
}
\description{
Adds a relativities step to a \code{rating_refinement} object.
}
\details{
\code{add_relativities()} is only available within the new refinement workflow
and should be used in combination with \code{prepare_refinement()} and \code{refit()}.
}
