% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_restriction}
\alias{add_restriction}
\alias{restrict_coef}
\title{Add coefficient restrictions to a refinement workflow}
\usage{
add_restriction(model, restrictions)

restrict_coef(model, restrictions)
}
\arguments{
\item{model}{Object of class \code{rating_refinement}.}

\item{restrictions}{data.frame with two columns containing restricted data.}
}
\value{
Object of class \code{rating_refinement}.
}
\description{
Adds a restriction step to a \code{rating_refinement} object.
}
\details{
\code{add_restriction()} is part of the new refinement workflow and is intended
to be used in combination with \code{prepare_refinement()} and \code{refit()}.

The legacy function \code{restrict_coef()} remains the only function that should
be applied directly to a model object.
}
