% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_smoothing}
\alias{add_smoothing}
\alias{smooth_coef}
\title{Add smoothing to a refinement workflow}
\usage{
add_smoothing(
  model,
  model_variable = NULL,
  source_variable = NULL,
  degree = NULL,
  breaks = NULL,
  smoothing = "spline",
  k = NULL,
  weights = NULL,
  tariff_class = NULL,
  rating_variable = NULL,
  x_cut = NULL,
  x_org = NULL
)

smooth_coef(
  model,
  x_cut,
  x_org,
  degree = NULL,
  breaks = NULL,
  smoothing = "spline",
  k = NULL,
  weights = NULL
)
}
\arguments{
\item{model}{Object of class \code{rating_refinement}, usually created with
\code{\link[=prepare_refinement]{prepare_refinement()}}.}

\item{model_variable}{Character string. Existing grouped or binned variable
in the GLM. This is the model term that will be replaced by a smoothed
tariff factor.}

\item{source_variable}{Character string. Original numeric portfolio variable
underlying \code{model_variable}.}

\item{degree}{Optional single whole number. Polynomial degree.}

\item{breaks}{Numeric vector with the new tariff class boundaries.}

\item{smoothing}{Character string with the smoothing method.}

\item{k}{Optional single positive whole number. Number of basis functions for
smoothing methods that use a basis dimension.}

\item{weights}{Optional character string. Weights column, usually exposure.}

\item{tariff_class, rating_variable}{Deprecated. Use \code{model_variable} and
\code{source_variable} instead.}

\item{x_cut, x_org}{Deprecated. Use \code{model_variable} and \code{source_variable}
instead.}
}
\value{
Object of class \code{rating_refinement}.
}
\description{
Adds a smoothing step to a \code{rating_refinement} object.

\code{smooth_coef()} is deprecated as of version 0.9.0.
Please use the refinement workflow instead:

\preformatted{
prepare_refinement(model) |>
  add_smoothing(...) |>
  refit()
}
}
\details{
\code{add_smoothing()} is part of the new refinement workflow and is intended
to be used in combination with \code{prepare_refinement()} and \code{refit()}.

The legacy function \code{smooth_coef()} remains the only function that should
be applied directly to a model object.
}
\seealso{
\code{\link[=add_smoothing]{add_smoothing()}}, \code{\link[=prepare_refinement]{prepare_refinement()}}, \code{\link[=refit]{refit()}}
}
