% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{add_smoothing}
\alias{add_smoothing}
\alias{smooth_coef}
\title{Add smoothing to a refinement workflow}
\usage{
add_smoothing(
  model,
  x_cut,
  x_org,
  degree = NULL,
  breaks = NULL,
  smoothing = "spline",
  k = NULL,
  weights = NULL
)

smooth_coef(
  model,
  x_cut,
  x_org,
  degree = NULL,
  breaks = NULL,
  smoothing = "spline",
  k = NULL,
  weights = NULL
)
}
\arguments{
\item{model}{Object of class \code{rating_refinement}.}

\item{x_cut}{column name with breaks/cut}

\item{x_org}{column name where x_cut is based on}

\item{degree}{order of polynomial}

\item{breaks}{numerical vector with new clusters for x}

\item{smoothing}{smoothing specification}

\item{k}{number of basis functions}

\item{weights}{weights column name, usually exposure}
}
\value{
Object of class \code{rating_refinement}.
}
\description{
Adds a smoothing step to a \code{rating_refinement} object.
}
\details{
\code{add_smoothing()} is part of the new refinement workflow and is intended
to be used in combination with \code{prepare_refinement()} and \code{refit()}.

The legacy function \code{smooth_coef()} remains the only function that should
be applied directly to a model object.
}
