% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_bootstrap_rmse.R
\name{autoplot.bootstrap_performance}
\alias{autoplot.bootstrap_performance}
\title{Autoplot for bootstrap_performance objects}
\usage{
\method{autoplot}{bootstrap_performance}(object, fill = "steelblue", color = "black", ...)
}
\arguments{
\item{object}{An object of class \code{"bootstrap_performance"}, produced by
\code{\link[=bootstrap_performance]{bootstrap_performance()}}.}

\item{fill}{Fill color of the histogram bars. Default = \code{"steelblue"}.}

\item{color}{Border color of the histogram bars. Default = \code{"black"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
\code{autoplot()} method for objects created by \code{\link[=bootstrap_performance]{bootstrap_performance()}}.
Produces a histogram and density plot of the bootstrapped RMSE values,
with the RMSE of the original fitted model shown as a dashed vertical line.
Optionally, 95\% quantile bounds are shown as dotted vertical lines.
}
\examples{
\dontrun{
mod1 <- glm(nclaims ~ age_policyholder, data = MTPL,
            offset = log(exposure), family = poisson())
x <- bootstrap_performance(mod1, MTPL, n = 100, show_progress = FALSE)
autoplot(x)
}

}
\author{
Martin Haringa
}
