% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rmse.R
\name{autoplot.bootstrap_rmse}
\alias{autoplot.bootstrap_rmse}
\title{Automatically create a ggplot for objects obtained from bootstrap_rmse()}
\usage{
\method{autoplot}{bootstrap_rmse}(object, fill = NULL, color = NULL, ...)
}
\arguments{
\item{object}{bootstrap_rmse object produced by \code{bootstrap_rmse()}}

\item{fill}{color to fill histogram (default is "steelblue")}

\item{color}{color to plot line colors of histogram}

\item{...}{other plotting parameters to affect the plot}
}
\value{
a ggplot object
}
\description{
Takes an object produced by \code{bootstrap_rmse()}, and plots the simulated RMSE
}
\author{
Martin Haringa
}
