% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{autoplot.check_residuals}
\alias{autoplot.check_residuals}
\title{Automatically create a ggplot for objects obtained from check_residuals()}
\usage{
\method{autoplot}{check_residuals}(object, show_message = TRUE, ...)
}
\arguments{
\item{object}{check_residuals object produced by \code{check_residuals()}}

\item{show_message}{show output from test (defaults to TRUE)}

\item{...}{other plotting parameters to affect the plot}
}
\value{
a ggplot object
}
\description{
Takes an object produced by \code{check_residuals()}, and produces a uniform quantile-quantile plot.
}
\author{
Martin Haringa
}
