% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{autoplot.restricted}
\alias{autoplot.restricted}
\title{Automatically create a ggplot for objects obtained from restrict_coef()}
\usage{
\method{autoplot}{restricted}(object, name = NULL, ...)
}
\arguments{
\item{object}{check_residuals object produced by \code{restrict_coef()}}

\item{name}{name of risk factor to show (defaults to NULL)}

\item{...}{other plotting parameters to affect the plot}
}
\value{
Object of class ggplot2
}
\description{
Takes an object produced by \code{restrict_coef()}, and produces
  a line plot with a comparison between the restricted coefficients and
  estimated coefficents obtained from the model.
}
\author{
Martin Haringa
}
