% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_severity.R
\name{autoplot.truncated_dist}
\alias{autoplot.truncated_dist}
\title{Create autoplot}
\usage{
\method{autoplot}{truncated_dist}(
  object,
  geom_ecdf = c("point", "step"),
  xlab = NULL,
  ylab = NULL,
  ylim = c(0, 1),
  xlim = NULL,
  print_title = TRUE,
  print_dig = 2,
  print_trunc = 2,
  ...
)
}
\arguments{
\item{object}{variable}

\item{geom_ecdf}{type of line}

\item{xlab}{label x}

\item{ylab}{label y}

\item{ylim}{y limit}

\item{xlim}{x limit}

\item{print_title}{true of false}

\item{print_dig}{number of digits}

\item{print_trunc}{number of truncation}

\item{...}{other plotting parameters to affect the plot}
}
\value{
ggplot2 object
}
\description{
Create autoplot
}
