% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_bootstrap_rmse.R
\name{bootstrap_rmse}
\alias{bootstrap_rmse}
\title{Bootstrapped RMSE}
\usage{
bootstrap_rmse(
  model,
  data,
  n = 50,
  frac = 1,
  show_progress = TRUE,
  rmse_model = NULL
)
}
\arguments{
\item{model}{A fitted model object.}

\item{data}{Data used to fit the model object.}

\item{n}{Integer. Number of bootstrap replicates. Default = 50.}

\item{frac}{Fraction of the data used in the training set if cross-validation
is applied. Must be in (0, 1]. Default = 1 (use all data).}

\item{show_progress}{Logical. Show progress bar during bootstrap iterations.
Default = TRUE.}

\item{rmse_model}{Optional numeric RMSE of the fitted (original) model.
If NULL (default), it is computed automatically.}
}
\value{
An object of class \code{"bootstrap_rmse"}, which is a list with components:
\describe{
\item{rmse_bs}{Numeric vector with \code{n} bootstrap RMSE values.}
\item{rmse_mod}{Root mean squared error for the original fitted model.}
}
}
\description{
Generate \code{n} bootstrap replicates to compute \code{n} root mean squared errors (RMSE).
This can be used to evaluate the predictive stability of a fitted model.
}
\details{
To test the predictive ability of a fitted model it can be helpful to assess
the variation in the computed RMSE. The variation is calculated by refitting
the model on \code{n} bootstrap replicates and storing the resulting RMSE values.
\itemize{
\item If \code{frac = 1}, each bootstrap sample has the same size as the dataset.
\item If \code{frac < 1}, a subset of the data is used as training, and the remainder
as test set (cross-validation).
}
}
\examples{
\dontrun{
mod1 <- glm(nclaims ~ age_policyholder, data = MTPL,
            offset = log(exposure), family = poisson())

# Use all records
x <- bootstrap_rmse(mod1, MTPL, n = 80, show_progress = FALSE)
print(x)
autoplot(x)

# Use 80\% of records (cross-validation style)
x_frac <- bootstrap_rmse(mod1, MTPL, n = 50, frac = .8, show_progress = FALSE)
autoplot(x_frac)
}

}
\author{
Martin Haringa
}
