% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_residuals.R
\name{check_residuals}
\alias{check_residuals}
\title{Check model residuals}
\usage{
check_residuals(object, n_simulations = 30)
}
\arguments{
\item{object}{A fitted model object (e.g. of class \code{"glm"}).}

\item{n_simulations}{Number of simulations to generate residuals. Default = 30.}
}
\value{
An object of class \code{"check_residuals"}, which is a list with:
\describe{
\item{df}{Data frame with theoretical quantiles (\code{x}) and observed residuals (\code{y}).}
\item{p.val}{P-value from Kolmogorov-Smirnov test against uniform(0,1).}
}
Invisibly returns the object.
}
\description{
Detect overall deviations of residuals from the expected distribution using
a simulation-based approach. Provides standardized residuals that are more
interpretable for GLMs than classical residual plots.
}
\details{
Misspecifications in GLMs cannot reliably be diagnosed with standard
residual plots, because the expected distribution of the data changes with
fitted values. \code{check_residuals()} uses a simulation-based approach
(similar to a parametric bootstrap or Bayesian p-value) to generate
standardized residuals between 0 and 1, which can be interpreted
intuitively like residuals from linear models.

This function wraps \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}, adapted for convenience.

Note: If all simulations for a data point have the same value (e.g. all
zeros), an error may occur (\verb{Error in approxfun: need at least two non-NA values}).
Increasing \code{n_simulations} can help in such cases.
}
\examples{
\dontrun{
m1 <- glm(nclaims ~ area, offset = log(exposure),
          family = poisson(), data = MTPL2)
cr <- check_residuals(m1, n_simulations = 50)
autoplot(cr)
}

}
\references{
Dunn, K. P., & Smyth, G. K. (1996). Randomized quantile residuals.
\emph{JCGS}, 5, 1–10.

Gelman, A., & Hill, J. (2006).
\emph{Data analysis using regression and multilevel/hierarchical models}.
Cambridge University Press.

Hartig, F. (2020). DHARMa: Residual Diagnostics for Hierarchical
(Multi-Level / Mixed) Regression Models. R package version 0.3.0.
\url{https://CRAN.R-project.org/package=DHARMa}
}
\author{
Martin Haringa
}
