% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_frequency.R
\name{clustering_frequency}
\alias{clustering_frequency}
\title{Breaks for continuous variables in insurance rating (frequency)}
\usage{
clustering_frequency(data, nclaims, x, exposure, cp = 0,
  color_splits = "red", color_gam = "steelblue", show_splits = TRUE)
}
\arguments{
\item{data}{dfd}

\item{nclaims}{ada}

\item{x}{ad}

\item{exposure}{adf}

\item{cp}{ddf}

\item{color_splits}{df}

\item{color_gam}{dfd}

\item{show_splits}{dfdf}
}
\value{
A list with components:

\item{x_clusters}{df}

\item{gam_plot}{}

\item{clusters}{Numerical vector with splits.}
}
\description{
The function provides an interface to finding class intervals for continuous numerical variables, for example to bin the continuous factors
such that categorical risk factors result which capture the effect of the covariate on the response in an accurate way,
while being easy to use in a generalized linear model (GLM).
}
\examples{
clustering_frequency(MTPL, nclaims, age_policyholder, exposure)
}
\author{
Martin Haringa
}
