% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_data.R
\name{construct_model_points}
\alias{construct_model_points}
\title{Construct model points from Generalized Linear Model}
\usage{
construct_model_points(
  x,
  expand = FALSE,
  drop_na = FALSE,
  exposure = NULL,
  exposure_by = NULL
)
}
\arguments{
\item{x}{Object of class refitsmooth or of class refitrestricted}

\item{expand}{expand true or false}

\item{drop_na}{drop na values}

\item{exposure}{column with exposure}

\item{exposure_by}{split column exposure by year column e.g.}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \code{construct_model_points()}
is used to construct model points from refitted generalized linear models,
and must be preceded by \code{update_glm()}.
}
\author{
Martin Haringa
}
