% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's natural breaks classification}
\usage{
fisher(vec, n = 7)
}
\arguments{
\item{n}{}
}
\value{
Vector with clustering
}
\description{
The functions provides an interface to finding class intervals for continuous numerical variables, for example for choosing colours for plotting.
}
\details{
The "fisher" style uses the algorithm proposed by W. D. Fisher (1958) and discussed by Slocum et al. (2005) as the Fisher-Jenks algorithm. This function is adopted from the classInt package.
}
\references{
Fisher, W. D. 1958 "On grouping for maximum homogeneity", Journal of the American Statistical Association, 53, pp. 789–798 (http://lib.stat.cmu.edu/cmlib/src/cluster/fish.f)
}
\author{
Martin Haringa
}
