% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_classify.R
\name{fisher_classify}
\alias{fisher_classify}
\alias{fisher}
\title{Fisher's natural breaks classification}
\usage{
fisher_classify(vec, n = 7, diglab = 2)

fisher(vec, n = 7, diglab = 2)
}
\arguments{
\item{vec}{A numeric vector to be classified.}

\item{n}{Integer. Number of classes to generate (default = 7).}

\item{diglab}{Integer. Number of significant digits to use for labels
(default = 2).}
}
\value{
A factor with class intervals as levels.
}
\description{
Classifies a continuous numeric vector into intervals using
Fisher-Jenks natural breaks. Useful for choropleth mapping or
other applications where grouped ranges are required.
}
\details{
The "fisher" style uses the algorithm proposed by W. D. Fisher (1958)
and discussed by Slocum et al. (2005) as the Fisher-Jenks algorithm.
This function is a wrapper around the \pkg{classInt} package.
}
\examples{
set.seed(1)
x <- rnorm(100)
fisher_classify(x, n = 5)

}
\references{
Bivand, R. (2018). \emph{classInt: Choose Univariate Class Intervals}.
R package version 0.2-3. \url{https://CRAN.R-project.org/package=classInt}

Fisher, W. D. (1958). \emph{On grouping for maximum homogeneity}.
Journal of the American Statistical Association, 53, pp. 789–798.
doi:10.1080/01621459.1958.10501479
}
\author{
Martin Haringa
}
