% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_reduce.R
\name{merge_date_ranges}
\alias{merge_date_ranges}
\alias{reduce}
\title{Reduce portfolio by merging redundant date ranges}
\usage{
merge_date_ranges(
  df,
  begin,
  end,
  ...,
  agg_cols = NULL,
  agg = "sum",
  min.gapwidth = 5
)

reduce(df, begin, end, ..., agg_cols = NULL, agg = "sum", min.gapwidth = 5)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{begin}{Name of column in \code{df} with begin dates.}

\item{end}{Name of column in \code{df} with end dates.}

\item{...}{Names of columns in \code{df} used to group date ranges by.}

\item{agg_cols}{List of columns in \code{df} to aggregate (default = NULL).}

\item{agg}{Aggregation function as character string (default = \code{"sum"}).}

\item{min.gapwidth}{Ranges separated by at least \code{min.gapwidth} days are not
merged (default = 5).}
}
\value{
A \code{data.table} of class \code{"reduce"}, with attributes:
\itemize{
\item \code{begin} — name of the begin-date column
\item \code{end}   — name of the end-date column
\item \code{cols}  — grouping columns
}
}
\description{
Transform all the date ranges together as a set to produce a
new set of date ranges. Ranges separated by a gap of at least \code{min.gapwidth}
days are not merged.
}
\details{
This function is inspired by \code{IRanges::reduce()}, adapted for
insurance portfolios.
}
\examples{
portfolio <- data.frame(
  policy_nr   = rep("12345", 11),
  productgroup= rep("fire", 11),
  product     = rep("contents", 11),
  begin_dat   = as.Date(c(16709,16740,16801,17410,17440,17805,17897,
                          17956,17987,18017,18262), origin="1970-01-01"),
  end_dat     = as.Date(c(16739,16800,16831,17439,17531,17896,17955,
                          17986,18016,18261,18292), origin="1970-01-01"),
  premium     = c(89,58,83,73,69,94,91,97,57,65,55)
)

# Merge periods
pt1 <- merge_date_ranges(portfolio, begin = begin_dat, end = end_dat,
    policy_nr, productgroup, product, min.gapwidth = 5)

# Aggregate per period
summary(pt1, period = "days", policy_nr, productgroup, product)

# Merge periods and sum premium per period
pt2 <- merge_date_ranges(portfolio, begin = begin_dat, end = end_dat,
    policy_nr, productgroup, product, agg_cols = list(premium),
    min.gapwidth = 5)

# Create summary with aggregation per week
summary(pt2, period = "weeks", policy_nr, productgroup, product)

}
\author{
Martin Haringa
}
