% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_data.R
\name{model_data}
\alias{model_data}
\title{Get model data}
\usage{
model_data(x)
}
\arguments{
\item{x}{Object of class refitsmooth or of class refitrestricted}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \code{construct_model_points()}
is used to construct model points from refitted generalized linear models,
and must be preceded by \code{update_glm()}.
}
\author{
Martin Haringa
}
