% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_data.R
\name{model_data}
\alias{model_data}
\title{Get model data}
\usage{
model_data(x)
}
\arguments{
\item{x}{Object of class refitsmooth, refitrestricted or glm}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \code{model_data()}
is used to get data from glm, and must be preceded by \code{update_glm()} or
\code{glm()}.
}
\author{
Martin Haringa
}
