% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPL.R
\docType{data}
\name{MTPL}
\alias{MTPL}
\title{Ages of 82,000 policyholders in a Motor Third Party Liability (MTPL) portfolio.}
\format{A data frame with 81830 rows and 3 variables:
\describe{
  \item{age_policyholder}{age of policyholder, in years}
  \item{nclaims}{number of claims}
  \item{exposure}{exposure, for example, if a vehicle is insured as of July 1 for a certain year, then during that year, this would represent an exposure of 1/2 to the insurance company.}
}}
\source{
The data is derived from the portfolio of a large Belgian motor insurance company.
}
\usage{
MTPL
}
\description{
A dataset containing the age, number of claims and exposure of almost 82,000 policyholders.
}
\keyword{datasets}
