% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_histogram.R
\name{outlier_histogram}
\alias{outlier_histogram}
\alias{histbin}
\title{Histogram with outlier bins}
\usage{
outlier_histogram(
  data,
  x,
  left = NULL,
  right = NULL,
  line = FALSE,
  bins = 30,
  fill = "steelblue",
  color = "white",
  fill_outliers = "#a7d1a7"
)

histbin(
  data,
  x,
  left = NULL,
  right = NULL,
  line = FALSE,
  bins = 30,
  fill = "steelblue",
  color = "white",
  fill_outliers = "#a7d1a7"
)
}
\arguments{
\item{data}{A data.frame containing the variable to plot.}

\item{x}{Variable name in \code{data} to map on the x-axis.}

\item{left}{Optional numeric, floor of the range. Values below are binned
together.}

\item{right}{Optional numeric, ceiling of the range. Values above are binned
together.}

\item{line}{Logical. If TRUE, add a density line. Default = FALSE.}

\item{bins}{Integer. Number of bins to use. Default = 30.}

\item{fill}{Fill color for bars. If NULL, a default is chosen.}

\item{color}{Line color for bars. If NULL, a default is chosen.}

\item{fill_outliers}{Fill color for outlier bins. Default = "#a7d1a7".}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Visualize the distribution of a continuous variable using bins.
Values below \code{left} or above \code{right} can be grouped into outlier bins
for compact display when the range of values is wide.

\code{\link[=histbin]{histbin()}} is deprecated as of version 0.8.0.
Please use \code{\link[=outlier_histogram]{outlier_histogram()}} instead.

In addition, note that \code{x} must now be passed as \strong{string}
(standard evaluation).
}
\details{
This is a wrapper around \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.
The method for handling outliers is based on
\url{https://edwinth.github.io/blog/outlier-bin/}.
}
\examples{
outlier_histogram(MTPL2, "premium")
outlier_histogram(MTPL2, "premium", left = 30, right = 120, bins = 30)

}
\author{
Martin Haringa
}
