% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_loss.R
\name{print.excess_threshold_assessment}
\alias{print.excess_threshold_assessment}
\title{Print an excess threshold assessment}
\usage{
\method{print}{excess_threshold_assessment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"excess_threshold_assessment"}.}

\item{...}{Reserved for future extensions.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Compact print method for objects returned by \code{\link[=assess_excess_thresholds]{assess_excess_thresholds()}}.
}
\author{
Martin Haringa
}
\keyword{internal}
