% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rating_table_deprecated.R
\name{rating_factors}
\alias{rating_factors}
\title{Deprecated alias for \code{rating_table()}}
\usage{
rating_factors(
  ...,
  model_data = NULL,
  exposure = TRUE,
  exposure_name = NULL,
  signif_stars = FALSE,
  exponentiate = TRUE,
  round_exposure = 0
)
}
\arguments{
\item{...}{glm object(s) produced by \code{glm()}, \code{refit()} or \code{refit_glm()}}

\item{model_data}{Optional data.frame used to create the model(s). If \code{NULL},
the function tries to use \code{model$data} for each supplied model.}

\item{exposure}{Logical or character. If \code{TRUE} (default), exposure is added
if it can be inferred from the model. If \code{FALSE}, no exposure is added.
If a character string is supplied, it is interpreted as the exposure column
name.}

\item{exposure_name}{Deprecated. Use \code{exposure_output} in \code{\link[=rating_table]{rating_table()}}
instead.}

\item{signif_stars}{Deprecated. Use \code{significance} in \code{\link[=rating_table]{rating_table()}}
instead.}

\item{exponentiate}{Logical. If \code{TRUE} (default), coefficients are
exponentiated and shown as relativities. If \code{FALSE}, coefficients are shown
on the model scale.}

\item{round_exposure}{number of digits for exposure (defaults to 0)}
}
\value{
See \code{\link[=rating_table]{rating_table()}}.
}
\description{
\code{rating_factors()} is deprecated as of version 0.8.0. Use
\code{\link[=rating_table]{rating_table()}} instead.
}
\keyword{internal}
