% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rating_factors.R
\name{rating_factors2}
\alias{rating_factors2}
\title{Include reference group in regression output}
\usage{
rating_factors2(
  model,
  model_data = NULL,
  exposure = NULL,
  colname = "estimate",
  exponentiate = TRUE,
  round_exposure = 0
)
}
\arguments{
\item{model}{glm object(s) produced by \code{glm()}}

\item{model_data}{data.frame used to create glm object(s), this should only
be specified in case the exposure is desired in the output, default value
is NULL.}

\item{exposure}{column in \code{model_data} with exposure, default value is NULL}

\item{colname}{name of column.}

\item{exponentiate}{logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to TRUE.}

\item{round_exposure}{number of digits for exposure (defaults to 0)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
