% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rating_table.R
\name{rating_table}
\alias{rating_table}
\alias{rating_factors}
\title{Include reference group in regression output}
\usage{
rating_table(
  ...,
  model_data = NULL,
  exposure = NULL,
  exponentiate = TRUE,
  signif_stars = FALSE,
  round_exposure = 0
)

rating_factors(
  ...,
  model_data = NULL,
  exposure = NULL,
  signif_stars = FALSE,
  exponentiate = TRUE,
  round_exposure = 0
)
}
\arguments{
\item{...}{glm object(s) produced by \code{glm()}, \code{refit()} or \code{refit_glm()}}

\item{model_data}{data.frame used to create glm object(s), this should only
be specified in case the exposure is desired in the output, default value
is NULL}

\item{exposure}{column in \code{model_data} with exposure, default value is NULL}

\item{exponentiate}{logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to TRUE.}

\item{signif_stars}{show significance stars for p-values (defaults to FALSE)}

\item{round_exposure}{number of digits for exposure (defaults to 0)}
}
\value{
Object of class \code{riskfactor}
}
\description{
Extract coefficients in terms of the original levels of the
coefficients rather than the coded variables.

\code{rating_table()} is intended for fitted models:
\itemize{
\item plain \code{glm} objects
\item models obtained after \code{refit()}
\item models obtained after \code{refit_glm()}
}

For pre-refit objects (\code{rating_refinement}, \code{restricted}, \code{smooth}) use
\code{print()}, \code{summary()} and \code{autoplot()} instead.
}
