% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit}
\alias{refit}
\title{Refit a prepared refinement workflow}
\usage{
refit(x, intercept_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{rating_refinement}.}

\item{intercept_only}{Logical. Default \code{FALSE}.}

\item{...}{Other arguments.}
}
\value{
Object of class \code{glm}.
}
\description{
Applies all collected refinement steps and refits the underlying GLM in one
call.
}
