% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit_glm}
\alias{refit_glm}
\alias{update_glm}
\title{Refit a GLM model or refinement workflow}
\usage{
refit_glm(x, intercept_only = FALSE, ...)

update_glm(x, intercept_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{rating_refinement}, \code{restricted} or \code{smooth}.}

\item{intercept_only}{Logical.}

\item{...}{Other arguments.}
}
\value{
Object of class \code{glm}.
}
\description{
Backwards-compatible wrapper. Prefer using \code{\link[=refit]{refit()}} for the new
refinement workflow.
}
