% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit_glm}
\alias{refit_glm}
\title{Refitting Generalized Linear Models}
\usage{
refit_glm(x)
}
\arguments{
\item{x}{Object of class restricted}
}
\value{
Object of class GLM
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
\code{refit_glm()} is used to refit generalized linear models, and must be
preceded by \code{restrict_coef()}.
}
\author{
Martin Haringa
}
