% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{refit_glm}
\alias{refit_glm}
\alias{update_glm}
\title{Refit a GLM model}
\usage{
refit_glm(x, intercept_only = FALSE, ...)

update_glm(x, intercept_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class restricted or of class smooth}

\item{intercept_only}{Logical. Default is \code{FALSE}. If \code{TRUE}, only
the intercept is updated, ensuring that the changes have no impact on the
other variables.}

\item{...}{Other arguments}
}
\value{
Object of class GLM
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This is the new version of \code{refit_glm()}. It refits the GLM with
any restrictions or smoothings that were previously added.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\author{
Martin Haringa
}
