% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_refinement.R
\name{relativities}
\alias{relativities}
\title{Combine multiple level splits into relativities}
\usage{
relativities(...)
}
\arguments{
\item{...}{One or more objects created by \code{split_level()}.}
}
\value{
A named list of data.frames suitable for the \code{relativities}
argument in \code{add_relativities()}.
}
\description{
Helper function to combine multiple level split definitions into a single
named list suitable for use in \code{add_relativities()}.
}
\examples{
relativities(
  split_level("construction",
              c("residential", "commercial", "civil"),
              c(1.00, 1.10, 1.25))
)

}
