% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{restrict_coef}
\alias{restrict_coef}
\title{Restrict coefficients in the model}
\usage{
restrict_coef(model, restrictions)
}
\arguments{
\item{model}{object of class glm/restricted}

\item{restrictions}{data.frame with two columns containing restricted data. The first
column, with the name of the risk factor as column name, must contain the
levels of the risk factor. The second column must contain the restricted
coefficients.}
}
\value{
Object of class restricted.
}
\description{
Add restrictions, like a bonus-malus structure, on the risk
factors used in the model. \code{restrict_coef()} must always be followed
by \code{refit_glm()}.
}
\details{
Although restrictions could be applied either to the frequency or
  the severity model, it is more appropriate to impose the restrictions
  on the premium model. This can be achieved by calculating the pure
  premium for each record (i.e. expected number of claims times the expected
  claim amount), then fitting an "unrestricted" Gamma GLM to the pure premium,
  and then imposing the restrictions in a final "restricted" Gamma GLM.
}
\author{
Martin Haringa
}
