% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{restrict_coef}
\alias{restrict_coef}
\title{Restrict coefficients}
\usage{
restrict_coef(model, data)
}
\arguments{
\item{model}{object of class glm/restricted}

\item{data}{data.frame with restricted data}
}
\value{
Object of class restricted
}
\description{
Restrict coefficients
}
