% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_severity.R
\name{rgammat}
\alias{rgammat}
\title{Generate data from truncated gamma distribution}
\usage{
rgammat(n, scale = scale, shape = shape, lower, upper)
}
\arguments{
\item{n}{number of observations}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{lower}{numeric. Observations below this threshold are not present in
the sample.}

\item{upper}{numeric. Observations above this threshold are not present in
the sample.}
}
\value{
The length of the result is determined by \code{n}.
}
\description{
Random generation for the truncated Gamma distribution with
parameters shape and scale.
}
\author{
Martin Haringa
}
