% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_severity.R
\name{rlnormt}
\alias{rlnormt}
\title{Generate data from truncated lognormal distribution}
\usage{
rlnormt(n, meanlog, sdlog, lower, upper)
}
\arguments{
\item{n}{number of observations}

\item{meanlog}{mean of the distribution on the log scale}

\item{sdlog}{standard deviation of the distribution on the log scale}

\item{lower}{numeric. Observations below this threshold are not present in the sample.}

\item{upper}{numeric. Observations above this threshold are not present in the sample.}
}
\value{
The length of the result is determined by \code{n}.
}
\description{
Random generation for the truncated log normal distribution whose logarithm has mean equal to meanlog and standard deviation equal to sdlog.
}
\author{
Martin Haringa
}
