% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_rows_per_date.R
\name{rows_per_date}
\alias{rows_per_date}
\title{Deprecated alias for \code{active_rows_by_date()}}
\usage{
rows_per_date(
  df,
  dates,
  df_begin,
  df_end,
  dates_date,
  ...,
  nomatch = NULL,
  mult = "all"
)
}
\arguments{
\item{df}{Deprecated. Use \code{portfolio} instead.}

\item{dates}{A \code{data.frame} or \code{data.table} with event or snapshot dates.}

\item{df_begin}{Deprecated NSE argument. Use \code{period_start} instead.}

\item{df_end}{Deprecated NSE argument. Use \code{period_end} instead.}

\item{dates_date}{Deprecated NSE argument. Use \code{date} instead.}

\item{...}{Deprecated NSE join columns. Use \code{by} instead.}

\item{nomatch}{When a row (with interval say, \verb{[a,b]}) in x has no match in
y, nomatch=NA means NA is returned for y's non-by.y columns for that row of
x. nomatch=NULL (default) means no rows will be returned for that row of x.}

\item{mult}{When multiple rows in y match to the row in x, \code{mult} controls
which values are returned - "all" (default), "first" or "last".}
}
\value{
See \code{\link[=active_rows_by_date]{active_rows_by_date()}}.
}
\description{
\code{rows_per_date()} is deprecated as of version 0.9.0. Use
\code{\link[=active_rows_by_date]{active_rows_by_date()}} instead.
}
\keyword{internal}
