% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_loss.R
\name{summary.excess_threshold_assessment}
\alias{summary.excess_threshold_assessment}
\title{Summarise an excess threshold assessment}
\usage{
\method{summary}{excess_threshold_assessment}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[=assess_excess_threshold]{assess_excess_threshold()}}.}

\item{...}{Unused.}
}
\value{
A \code{data.frame}.
}
\description{
Return the key threshold diagnostics from an
\code{"excess_threshold_assessment"} object.
}
\author{
Martin Haringa
}
\keyword{internal}
