% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_reduce.R
\name{summary.reduce}
\alias{summary.reduce}
\title{Automatically create a summary for objects obtained from reduce()}
\usage{
\method{summary}{reduce}(object, ..., period = "days", name = "count")
}
\arguments{
\item{object}{reduce object produced by \code{reduce()}}

\item{...}{names of columns to aggregate counts by}

\item{period}{a character string indicating the period to aggregate on.
Four options are available: "quarters", "months", "weeks", and "days"
(the default option)}

\item{name}{The name of the new column in the output. If omitted, it will
default to count.}
}
\value{
data.frame
}
\description{
Takes an object produced by \code{reduce()}, and counts new and lost
customers.
}
