% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_refinement.R
\name{update_glm}
\alias{update_glm}
\title{Refitting Generalized Linear Models}
\usage{
update_glm(x)
}
\arguments{
\item{x}{Object of class restricted or of class smooth}
}
\value{
Object of class GLM
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{update_glm()} is used to refit generalized linear models, and must be
preceded by \code{restrict_coef()}.
}
\author{
Martin Haringa
}
